Escalada plugin for GLPI V 2.0.2
glpi-escalade-2.0.2.tar.gz  GLPI 0.84  Miércoles, 12 de Marzo de 2014
15:36:06 +0100
climbing

This extension simplifies the process of climbing back and between
groups within a ticket.
It adds a visual history of the groups involved in the ticket, a new
dashboard on the center page and a new search criterion for tickets.
Warning: it disables the ability to assign multiple groups to a single
ticket (applicants and observers are not affected)