Helpdeskrating plugin for GLPI V 1.1.0
helpdeskrating_1.1.0.zip Last Update: 2013-08-10 0.84.X
Helpdeskrating is a Plugin for GLPI http://www.glpi-project.org
extending the helpdesk tickets with a new tab. Here the work of the
technician can be rated in 4 different
categories (overall, sollution, technician, time) with points from 1 to
6, where 1 is best and 6 is worst.

In turn the technician can also rate the requester's communication,
input and collaboration, also with points from 1 to 6.

Additional to that the requester and the technician have the
possibillity to add a comment to their rating, to explain the given
points, if needed.

The rating is possible as soon as the ticket is solved.

GLPI-Logo by the GLPI-Project
Thumbs in Logo by http://commons.wikimedia.org/wiki/User:Bersam

Rating of Helpdesk tickets by the requester and the technician
Rating is optional and always possible as soon the ticket is solved
Rating a Ticket automatically accepts the solution
Different statistics for the technician