Winadminpassword plugin for GLPI V 1.1.1
glpi-winadminpassword-1.1.1.tar.gz  Compatible GLPI >= 0.84.x and < 0.85
Lunes, 30 de Setiembre de 2013 22:45:35 +0200
GLPI plugin for WinAdminPassword tool.
WinAdminPassword tool :
[[http://sourceforge.net/projects/winadminpasswd/]]
Tool for the deployment of unique passwords for Windows and Unix
systems. It is based on the serial number of computers and a secret key.
The advantage is that no password is stored in a database and you can
display them with a GLPI plugin, Webmin...
Generate passwords based on a secret key and computers serial numbers
Generate passwords based on the system time
Change user passwords on Microsoft Windows and Unix systems
Print serial number on Microsoft Windows and Unix systems
display generated passwords
Add a system task to change password periodically with a password based
on the system time